requirements.txt
requests
python-dotenv
praw
lyricsgenius
